requirements.txt
requests
paho-mqtt
configparser
pytz
